backend/requirements/requirements.txt
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
boto3==1.37.20
botocore==1.37.20
Django==5.1.7
django-cors-headers==4.6.0
djangorestframework==3.16.0
dnspython==2.7.0
docker==7.1.0
Flask-WTF==1.2.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
gunicorn==23.0.0
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
starlette==0.46.1

